// src/_services/api/model/terminoCarregamentoGetDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API de Gestão de Devolução e Produtividade
 * Documentação da API responsável por: gestão de produtividade, montagem e impressão de mapas de separação, sistema de devolução de estoque e contagem de inventário.
 * OpenAPI spec version: 1.0
 */
import type { TerminoCarregamentoGetDtoHorarioTerminoCarregamento } from './terminoCarregamentoGetDtoHorarioTerminoCarregamento';
import type { TerminoCarregamentoGetDtoNumeroTransporte } from './terminoCarregamentoGetDtoNumeroTransporte';
import type { TerminoCarregamentoGetDtoDataExpedicao } from './terminoCarregamentoGetDtoDataExpedicao';
import type { TerminoCarregamentoGetDtoCenterId } from './terminoCarregamentoGetDtoCenterId';

export interface TerminoCarregamentoGetDto {
  horarioTerminoCarregamento: TerminoCarregamentoGetDtoHorarioTerminoCarregamento;
  numeroTransporte: TerminoCarregamentoGetDtoNumeroTransporte;
  dataExpedicao: TerminoCarregamentoGetDtoDataExpedicao;
  centerId: TerminoCarregamentoGetDtoCenterId;
}